# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Millionaire Quiz API",
        "description": "API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/multiplayer/battle/join/{inviteCode}": {
            "post": {
                "tags": [
                    "Multiplayer"
                ],
                "summary": "Join a battle lobby",
                "operationId": "d3a1c6996e869dedbe28face740051f8",
                "responses": {
                    "200": {
                        "description": "Joined lobby successfully"
                    },
                    "400": {
                        "description": "Failed to join lobby"
                    }
                }
            }
        },
        "/multiplayer/battle/lobby/{inviteCode}": {
            "get": {
                "tags": [
                    "Multiplayer"
                ],
                "summary": "Get battle lobby details",
                "operationId": "2f97de207eb69341373d842648d57546",
                "responses": {
                    "200": {
                        "description": "Lobby details retrieved"
                    },
                    "404": {
                        "description": "Lobby not found"
                    }
                }
            }
        },
...
        "/multiplayer/battle/start/{inviteCode}": {
            "post": {
                "tags": [
                    "Multiplayer"
                ],
                "summary": "Start the battle (host only)",
                "operationId": "84fa31091d24a5c6a5c11b860c726585",
                "responses": {
                    "200": {
                        "description": "Battle started successfully"
                    },
                    "400": {
                        "description": "Failed to start battle"
                    }
                }
            }
        },
...
        "/multiplayer/action": {
            "post": {
                "tags": [
                    "Multiplayer"
                ],
                "summary": "Send game action to other players",
                "operationId": "553b8db3e76fc035f0f4eb87886d7182",
                "responses": {
                    "200": {
                        "description": "Action broadcasted successfully"
                    },
                    "500": {
                        "description": "Failed to broadcast action"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8001/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8001/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8001/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}